Fix to 'should_generate' which unnecessarily retrieves all project document in command-line mode
git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@42145 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/tools/doc_builder/xml/xslt/html_generator.e
@@ -50,18 +50,31 @@ feature -- Generation
 			l_src,
 			l_target: DIRECTORY
 		do
+			print ("Starting HTML generation...%N")
+			print ("Making src directory (" + Shared_project.root_directory + ")%N")
 			create l_src.make (Shared_project.root_directory)
+			print ("Making target directory (" + location + ")%N")
 			create l_target.make (location)
 			if not l_target.exists then
+				print ("Creating target directory since it does not exist%N")
 				l_target.create_dir
 			end
+			print ("Deleting contents of target directory%N")
 			l_target.delete_content
 			if should_generate then
+				print ("Entering generation%N")
+				print ("Setting title%N")
 				progress_generator.set_title ("HTML Generation")
+				print ("Setting procedure%N")
 				progress_generator.set_procedure (agent generate_directory (l_src, l_target))
+				print ("Setting upper range%N")
 				progress_generator.set_upper_range (files.count)
+				print ("Setting heading text%N")
 				progress_generator.set_heading_text ("Generating HTML Files...")
+				print ("Generating%N")
 				progress_generator.generate
+			else
+				print ("Not generating%N")
 			end
 		end	
 
@@ -237,18 +250,16 @@ feature {NONE} -- Implementation
 			l_question_dialog: EV_MESSAGE_DIALOG
 		do
 			Result := True
-			if Shared_project.has_invalid_files then
-				if Shared_constants.Application_constants.is_gui_mode then
-					create l_question_dialog.make_with_text ((create {MESSAGE_CONSTANTS}).invalid_project_files_warning)
-					l_question_dialog.set_title ((create {MESSAGE_CONSTANTS}).report_title)
-					l_question_dialog.set_buttons (<<"Continue", (create {EV_DIALOG_CONSTANTS}).ev_cancel>>)
-					l_question_dialog.show_modal_to_window (Application_window)
-					if l_question_dialog.selected_button.is_equal ("Continue") then
-						l_question_dialog.destroy
-					else
-						l_question_dialog.destroy
-						Result := False
-					end				
+			if shared_constants.application_constants.is_gui_mode and then Shared_project.has_invalid_files then
+				create l_question_dialog.make_with_text ((create {MESSAGE_CONSTANTS}).invalid_project_files_warning)
+				l_question_dialog.set_title ((create {MESSAGE_CONSTANTS}).report_title)
+				l_question_dialog.set_buttons (<<"Continue", (create {EV_DIALOG_CONSTANTS}).ev_cancel>>)
+				l_question_dialog.show_modal_to_window (Application_window)
+				if l_question_dialog.selected_button.is_equal ("Continue") then
+					l_question_dialog.destroy
+				else
+					l_question_dialog.destroy
+					Result := False
 				end				
 			end
 		end
